Extract code for file resolving from bpmn preview
